Use upstream ref if possible
When the current branch tracks an upstream branch, generate a link to
that upstream branch (or SHA) rather than using the local one, since it
may not exist upstream.

In the case that the local branch has no upstream, prompt the user for
the branch name to generate a link for.

Partly addresses https://github.com/ruanyl/vim-gh-line/issues/10.

Co-authored-by: souldzin <souldzin@souldzin.com>